requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
reportlab==4.4.7
sqlparse==0.5.5
